moved files around to resemble the new file structure
