+++ src/partials/Footer/index.ts
